MusicBandSupporter-C#
This is TestMusicBandSupporter, an introductory exercise to learn about test classes and working with objects from another class. This is the rewrite in C# for personal learning. In doing this I learned about class/namespace references in C# (no reference needed in Java) and I learned that default visibility does not allow access even in the same package/solution in C#.